requirements.txt
python-telegram-bot[job-queue]>=21.6
aiohttp
python-dotenv
tzdata
beautifulsoup4
lxml
